requirements.txt
python-telegram-bot==20.7
requests==2.31.0
aiofiles==23.2.1
aiohttp==3.9.1
python-dotenv==1.0.0
Pillow==10.1.0
tqdm==4.66.1
humanize==4.8.0
zipfile36==0.1.3
psutil==5.9.6